backend/requirements.txt
flask
sqlalchemy
Flask-Login
Flask-Bcrypt
Flask-SQLAlchemy
Flask-Cors
python-dotenv
